Extract Schema class
